refactor(utils): extract detectQuoteZones, citationBounds, parenDepths
Three mature private helpers moved to src/utils/ for reuse by the
upcoming analyzeDocument module:

- detectQuoteZones (was inline in DocumentResolver.ts)
- getCitationStart/End (was inline in detectStringCites.ts)
- computeParenDepths (was a private method on DocumentResolver)

Algorithms unchanged. Adds focused unit tests for each — previously
only tested indirectly through their callers. DocumentResolver and
detectStringCites updated to import from utils/.
